Add hover enable/disable checkboxes to lesHoverables autotest qml
This isn't needed for the autotest, but we sometimes use this qml file
for manual testing.

Control-h toggles the upper HoverHandler enabled state, and Control-m
toggles the MouseArea hoverEnabled state; that way it's possible to
test that hover feedback disappears even if the mouse isn't moved.
However, flushFrameSynchronousEvents() currently gets called only
if some items are marked dirty; so if we don't change the checkbox
appearance, _and_ we use only keyboard shortcuts to change the
enabled state, this demonstrates that we still have bugs: we don't
immediately update the containsMouse and hovered properties when
the respective objects have their hover processing en-/disabled.

While we're at it: stop using Loader, use named components instead.

Task-number: QTBUG-46460
Change-Id: Ib7aecf4788621a7d5cf0b9461ca56553610120ec
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>