qml: allow inline component types as signal argument
The resolution of inline components through QQmlType was failing and
therefore inline components were not allowed as signal parameters.

Apply the same fix as 2a37ff2f49140272d0122ccc097cc14c2fa4133e to signal
parameter-type resolution.

Test if signals with inline-component-typed parameters works, even if
the inline component is defined in another qml file.

SignalInlineComponentArg.qml is capitalized as it is used in
signalInlineComponentArg1.qml.

Pick-to: 6.4 6.3 6.2
Fixes: QTBUG-106611
Change-Id: I2bbcee56025e6a319a3fea9b7aedf703afabe6b3
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>