QuickControls2: implement icons for submenus
Previously in Qt 5 there was an option to define iconSource property
in Controls 1 Menu. But it's missed in Controls 2, so it's not possible
to use icon for submenu item in Qt 6. This patch introduces icon group
property for the Menu so that it behaves and displayed the same as for
regular menu items.

[ChangeLog][QtQuick][Controls] Now it's possible to set icon for Menu.
This is useful for submenus, so icon will be displayed as for regular
menu items.

Fixes: QTBUG-79972
Change-Id: I4610d4b6d40c51b774b035e238e5e1ceea53a8a3
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>