Dialogs: Defer automatic window resolving until dialog is opened
The qml tool, as well as the QQuickView class are creating their own
window outside of the root qml file that is passed to them.

This caused an issue where no dialog were able to find a window when
componentComplete() was first called, which would make them return
early in the QPlatformDialogHelper::show() re-implementations, and
thus never show the dialogs when open() was called.

This patch solves the issue by waiting until open() is called, before
trying to find a window in any of the parent items.

Since Window always has a contentItem, it was also possible to simplify
the logic that searches for a window, by only trying to cast the parent
to QQuickItem*.

Another issue that came up, was the fact that dialogs with visible set
to true, aka:

FileDialog {
    visible: true
}

Should open as soon as possible upon its creation.
To get around this, open() is called during componentComplete() if a
window can be found during this time. Otherwise, the call to open() is
deferred until the nearest parent item emits the windowChanged() signal.

Note that calling setParentWindow() will now take priority over the
implicit window search, unless the argument for setParentWindow is
null. Meaning that moving a dialog to a different window, by modifying
parents, will not happen for dialogs that are explicitly binding a value
to parentWindow.

Fixes: QTBUG-106598
Pick-to: 6.5
Change-Id: Idd7c0ecdeea6cbf26e8d41168788563ee9794118
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>
(cherry picked from commit f233a5ff9d04daf3ba792cc196da7e5f190b415c)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>
(cherry picked from commit ba596daec67fb41d87d6c4a0b31f215fc0a1359d)