QQmlThread: get rid of startup synchronization
startup() waits for a notification from the worker thread, which does
this as its first action in run(). However, there is no data set up by
the worker thread that would require this kind of synchronization.
As the whole procedure is superfluous, simplify startup to only start
the QThread, and move it to itself (an anti-pattern to be revisited
later).
We also can get rid of the run() override, as the default implementation
does everything we need - namely starting the event loop.

Change-Id: I648778913893ebce9efc6aa61e38ed38cad0b0e7
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>