Merge QQuickShaderEffectImpl into Private
With qmlbench this gives an improvement (405->430) in
delegates_shadereffect.qml.

QQuickShaderEffectImpl is not needed anymore in Qt 6, because it
only has one implementation (unlike in Qt 5.14 and 5.15). Move it
all into QQuickShaderEffectPrivate.

Fixes an obvious bug as well: in propertyChanged the derefWindow()
call should be called with the old 'source' object, not the new one.

Change-Id: I13d2a3a9134368e3319e3967bfb994eb92866704
Reviewed-by: Andy Nichols <andy.nichols@qt.io>