Don't used threaded pixmap downloads on WASM
WASM sockets rely on async callbacks that won't arrive when a thread is
waiting on a condition in QEventDispatcherWasm::wait.
The thread deadlocks, as the condition will not ever be signaled, since
that relies on the async callback from emscripten_fetch being issued.

Because WASM networking is non-blocking anyway, we forgo using the
threaded download altogether.

Change-Id: Ic75bcf7dd6272a6d3ea7b934807fcc74458b13ba
Fixes: QTBUG-106159
Reviewed-by: Morten Johan Sørvig <morten.sorvig@qt.io>