DA: don't exit early from deliverHoverEvents when we have hovered items
The current implementation would exit early if subtreeHoverEnabled
was false. The problem is that this can also happen if you set
hoverEnabled to false while an item is being hovered. And in that
case, we need to prune the list of hovered items.

This patch will therefore ensure that we only exit early if
subtreeHoverEnabled is false, _and_ no items are currently
marked as being hovered.

Change-Id: I6086c7cbea07b0973db343a3f439810361c5d1b7
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>
(cherry picked from commit 255c04132455ff404245e9872ef868db528a0bcb)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>