QmlObjectCreator: play nice with grouped properties
If an object tries to set up a grouped property binding of the form
complex.subprop: primitiveValue
when complex is declared in a base element, and the setup of it is done
in base, then we so far would get an error about complex being null.

The reason for that is an optimization which only sets up bindings
shortly before object creation concludes (to avoid spurious
reevaluations during construction). Because setting primitive value
"bindings" happens however without delay, we end up with the case of
attempting to set up the value of a non-existent object.

To fix this we check if there is a binding which might provide the
object, evaluate it at that point, and to then re-read the outer
property of the group.

As a drive-by, add missing QT_{BEGIN,END}_NAMESPACE macro to
qqmlanybinding_p.h. Otherwise, the namespace build (finally) fails to
compile.

Pick-to: 6.5
Fixes: QTBUG-108102
Change-Id: Ifdac1948100b369a49c0a5ecff75f881fd63540e
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>