QQuickTableView: add API to set section sizes explicitly
Add API in TableView for setting column and row sizes
explicitly.

Up till now, the recommended way to set column widths
in TableView, was to use a columnWidthProvider (or
simply rely on the implicit width of the delegates).
The reason is that this scales much better, compared
to imperatively setting all the column widths up-front,
e.g at application start-up.

But there are several cases where this API becomes a bit
cumbersome to use, epecially for smaller data models, or
when the rows and columns should be allowed to be resized
at run-time. One such usecase is the ongoing work of
supporting section resizing in HeaderView.

This patch will therefore add an API that allows
HeaderView, as well as the application, to set column
and row sizes explicitly. More documentation on how
this works is included in the patch.

[ChangeLog][Quick][TableView] Additional API for setting
explicit row-, and column sizes has been made available.

Task-number: QTBUG-106792
Change-Id: I1cbcd36a57470580b2548442275a818a837ea2be
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>