sg: Limit automatic pipeline cache writes
...based on the window flags.

As the docs already describe, for full control one can assign cache
files explicitly to each QQuickWindow. (of course, that is not always
trivial esp. when working with the QML API but let's not go there now)
This does not affect that. The default, automatic cache however needs
to apply some simple guesses as to when it does not make sense to
bother with loading/saving the cache. Otherwise we'd end up with
silly situations, when using the default threaded render loop at
least, where showing/destroying a tooltip (if it's a standalone
window) kicks off file I/O for the pipeline cache, with likely
very limited benefits (and whatever it collects is lost later
when the main (last) window is closed anyway).

In addition, make some of the logs more readable by including
which QRhi and which QQuickWindow are the messages talking about.
This becomes pretty important in applications dealing with
multiple windows over their lifetime.

Change-Id: I9cabe3eb0db994e5c1d73ca071664ff1bf5f0d45
Reviewed-by: Christian Strømme <christian.stromme@qt.io>