Ensure that item appearance is updated when hover state changes
If a HoverHandler is enabled/disabled, or a MouseArea.hoverEnabled is
set or unset, its parent may need to change its hovered appearance.
Marking an item dirty is enough to ensure that
flushFrameSynchronousEvents() will be called, in case there was no
other event or property change to trigger that.

Pick-to: 6.4
Task-number: QTBUG-46460
Fixes: QTBUG-106704
Change-Id: I5ced5f673baf387bbd0ee678e120538d1bb61c0e
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>