DA: Avoid corrupting persistent scenePosition in QHoverEvent ctor
It's important to set lastMousePosition with the correct scene position
in QQuickDeliveryAgent::event's QEvent::Enter case, not some
item-localized position.

The QHoverEvent ctor takes two points: pos and globalPos; pos is then
passed as both the scene and global pos to the QSinglePointEvent ctor,
which calls QMutableEventPoint::setScenePosition() on the persistent
QEventPoint instance and then detaches befeore setting ephemeral state.
Therefore, we must construct QHoverEvent with scene position first, not
local position, so that the right value is persisted (or actually left
unchanged in this case); it's better to set local position after the
detach(), whereas it's too late to fix the persistent point then.

Pick-to: 6.4 6.2
Fixes: QTBUG-35409
Change-Id: Ibe840f8ac7b91fd711fc0f4d4a90a24e0117791b
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>