Avoid context truncation when filename has multiple dots
Use completeBaseName() as in lupdate while computing context. This
change causes regression as one of the test case relies on truncation of
the context. In order to prevent this from failing, introduce the file that test
case needs as input. Also, add a test case for not truncating multiple
dotted input file.

[ChangeLog][QtQml] The translation context in QML files has been changed
to be in line with lupdate: Instead of using only the baseName(),
completeBaseName() is now used.

Fixes: QTBUG-106705
Change-Id: Iae52657f28429109b4f32655cc95f444de72b0ba
Reviewed-by: Jörg Bornemann <joerg.bornemann@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>