createTemporaryObject: account for undefined parent argument
c837be7beab1c217b8f163b9a2d53ca12fd2c95e results in a warning for code
that does this:

    let parent = createTemporaryObject(window)

The warning was:

"QML Component: Unsuitable arguments passed to createObject(). The first
argument should be a QObject* or null, and the second argument should
be a JavaScript object or a QVariantMap"

Avoid the warning by setting parent to null if it's undefined.

Update the relevant self-tests to fail on warnings so that it's
tested.

Change-Id: I04e9423e7c08d370d8055a8b8e766827a9709919
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
(cherry picked from commit e7f23ef9b09b5fc522598abec20e541490ceba3d)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>