sg: Fix QQGraphicsConfig flag handling and add a test
The bitmask values were wrong which led to flags not set/unset
possibly reported with the opposite value.

Add support for qDebug printing. This we will use in QSGRhiSupport
to improve logging but that needs to be a separate patch due to
some conflicts.

Add a minimal autotest as well.

Change-Id: Ib64f9f854e50224ec5f22b365c36b86208df9336
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Christian Strømme <christian.stromme@qt.io>