Make basic render loop use a dedicated QRhi per window
[ChangeLog][QtQuick] The 'basic' render loop has been changed
to use a dedicated graphics context/device for each window.
This is identical to the commonly used 'threaded' render
loop's behavior. This should not lead to any changes in
rendering or visual appearance, but is expected to
correct a number of lower level issues that tend to
arise due to not having control over the context/device
on a per-window basis.

Fixes: QTBUG-80508
Change-Id: Ia47085005871cfd648262f1c72cc9024298b1ff0
Reviewed-by: Andy Nichols <andy.nichols@qt.io>