CMake: Pass --private-includes to qmltyperegistrar for *Private targets
If the QML module is attached to the private target itself we want to
generate 'private/' file names, too. So far we only do that when the
QML module was attached to a public target that had a corresponding
private target.

Fixes: QTBUG-106887
Change-Id: I729ce39f362b7bdae02e4d737a3fc79e25c02ac1
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
(cherry picked from commit a5a7407b6bfa0ab07cc031740726dcc9fc234e72)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>