qml: deprecate top-level components
Print a deprecation warning during the IR compilation in the
qqmltypecompiler.cpp when a top-level capital-c-Component is found.
Also fix all the tests in tst_qqmllanguage that use a top level
component (either by wrapping those Components into Items or by removing
the explicit redundant Component-wrapping). This implies also fixing
the line numbers for Components in tests that got wrapped into an Item.

[ChangeLog][QtQml][QQmlComponent] QQmlComponents in qml document roots
or in inline component roots are deprecated. Types defined in qml
documents or in inline components are automatically wrapped into
QQmlComponents when needed: those types really should not be inside a
Component{}.

Fixes: QTBUG-106547
Change-Id: I00630b879f52f681c98a1d4b804e3505f175b9d6
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>