V4: Optimize more type conversions
When loading from value type properties we left quite a few trivial
conversions on the table and went via QVariant instead. Add them.

It would be nice to unify the various places where we switch over the
trivial types, but I don't see an easy way to do so right now.

Change-Id: I9dcb29ce147a2f282e7ea6903a7af8a4b3038af6
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>