Popup: Don't propagate unhandled mouse events beyond the overlay
After f0b278c4d0c4b50fc6947b4b1beb07ae172546c6, mouse events delivered
to a popup are not accepted by the popup, allowing children of the popup
(such as a mouse area) to handle the event. The assumption was then that
the overlay would ultimately handle and eat the event in its event
filter when the event delivery agent returns after handling the event.

However, the delivery agent delivers input events to all items in the
scene, not just to items "above" the overlay. Items under the popup
got the event as well, causing a regression allowing interaction with
items under a popup.

We still must not block events delivered to the popup to allow
propagation to child items. If not handled by any child, then the event
will ultimately propagate to the overlay item itself. The overlay
handles the event by giving the popup a chance to block the event,
passing itself as the source item. This allows e.g. a modal popup to
block a press outside the popup item hierarchy, by blockInput returning
true.

We also need to return true if the item passed in is the overlay, and if
the event is within the popup item. We know at this point that no item
within the popup has accepted the event, and must block the event from
propagating beyond the overlay.

Fixes: QTBUG-107014
Change-Id: Ie46799cafbd812fabd76c8b47ef270f43cbd563f
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
(cherry picked from commit e4a159632ec39542ec23567963191162f9dcdfef)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>