Doc: fix FrameAnimation's frameTime snippet
Add to rotation instead of assigning so that the Rectangle actually
rotates.

While we're at it, slow down the rotation so you can see that it's
smooth and center the Rectangle so its edges aren't out of view when
it rotates.

Change-Id: I126b00e4b28b9ee79889d0b5c274c9f0ee58a3ec
Reviewed-by: Kaj Grönholm <kaj.gronholm@qt.io>
(cherry picked from commit 30104dabe6cf73d8749537aa4b537b11dad26b40)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>