qmltc: fix code generation for implicit components
Do not check if a type is implictly wrapped during visitation, as this
information is only made available by QQmlJSImportVisitor after(!) the
visit. Move this into the postvisit step of qmltc, and write some
internal documentation so this error will hopefully not be done again by
some qmltc contributor (e.g., me).

Now that implicit components are correctly recognized as such, they can
be compiled correctly and the c++-code generated by qmltc can again be
compiled.

Added also a test to see if all childrens of the repeater are correctly
instantiated.

Pick-to: 6.4
Fixes: QTBUG-107091
Change-Id: I8fce018b83316f8786ae5ca15e5af27c30bb1d37
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>