Flickable: Proportional wheel scrolling if deceleration is large
Many users have complained about scrolling via mouse wheel being more or
less uncontrollable: it's very hard to scroll shorter distances, and
multiple "clicks" of the wheel usually overshoot far beyond what was
intended. It was possible to mitigate that by setting flickDeceleration,
but then touch-flicking also became much more "sticky". So now we have
the flickDeceleration property dedicated to touchscreen flicks only,
with the default coming from QPlatformIntegration::FlickDeceleration.
There's no QPlatformIntegration::StyleHint for mouse-wheel deceleration.
We disable the old wheel-acceleration feature by default, because many
users nowadays have fancier mice or touchpads that achieve the
acceleration in their own way. That's achieved by having the wheel
deceleration value set to 15000 by default, which would be quite "stiff"
if it was applied to the timeline, but in fact is defined to mean that
there will be no acceleration. If the default is changed to 14999 or
less, we continue reusing the same animation timeline for both flicking
and wheeling, and applying the same acceleration/deceleration as before.

There's some risk that some users will miss the old behavior; so we add
a new QT_QUICK_FLICKABLE_WHEEL_DECELERATION environment variable to
allow the user to override the default (to something less than 15000).

When acceleration is disabled (which it is by default),
QQuickFlickable::wheelEvent() no longer calls flick(), but instead uses
the timeline to move the contentItem directly.

Includes a drive-by fix, that enables scrolling to continue,
even after the contentItem has moved into an edge.
This behavioral change makes sense now that the scroll movement
is more predictable. fixup() will still be called,
to trigger the rebound animation.

This patch reduces the strength of wheelEvent scrolling, so it
made small tweeks to tst_QQuickListView2::wheelSnap necessary.

[ChangeLog][QtQuick][Flickable] When using a plain "clicky" mouse wheel,
Flickable has historically applied acceleration: the faster you rotate
the wheel, the faster it flicked. The default value of flickDeceleration
was 5000, which felt hard to control, but changing it in QML affected
touch-flicking and wheel behavior simultaneously. So now we decouple
those: flickDeceleration only affects touchscreen flicking, and
Flickable rather uses a new default value of 15000, which is defined
to mean that scroll distance becomes directly proportional to
QWheelEvent::angleDelta() * QStyleHints::wheelScrollLines() (where a
"line" is 24 logical pixels, because Flickable does not have direct
understanding of its contents). You can set the new environment var
QT_QUICK_FLICKABLE_WHEEL_DECELERATION to a value from 1 to 14999 to
restore wheel acceleration behavior from older versions; 5000 was
the old value.

Done-with: Oliver Eftevaag <oliver.eftevaag@qt.io>
Done-with: Paul Olav Tvete <paul.tvete@qt.io>
Task-number: QTBUG-35609
Task-number: QTBUG-80720
Task-number: QTBUG-82565
Fixes: QTBUG-97055
Fixes: QTBUG-106338
Pick-to: 6.6
Change-Id: I0645c08c9681f57bc749b4cc7f4d26507f5cb211
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Oliver Eftevaag <oliver.eftevaag@qt.io>