Locale: add toString functions
Currently, it's not possible to show a single localized day number;
Qt.formatDate can show it if "d" is passed as the format string, but
that rules out passing a locale. If a locale is passed instead, you
can only pass e.g. Locale.ShortFormat or another enum value as the
extra argument, as opposed to a format string. This limitation requires
users to implement conversion helpers in C++ that allow accounting
for the locale.

This patch exposes an appropriate subset of the existing date-to-string
conversion functions from QLocale to the QML Locale type.
For completeness, it also exposes the other relevant overloads:

- toString(int i)
- toString(double f, char format = 'g', int precision = 6)
- toString(Date date, string format)
- toString(Date date, FormatType format = LongFormat)

[ChangeLog][QtQml][Locale] Added toString() overloads.

Fixes: QTBUG-105506
Change-Id: I7061c68097cc11b69179117f5825e368bcfc296b
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Nicholas Bennett <nicholas.bennett@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>