Android: Set EnterKeyNext as default type for TextInput
Behavior of EnterKey for virtual keyboard need to be changed for
TextInput. Before this commit, ImeOption was set to IME_ACTION_DONE.
Beacuse of that, setting any text in TextInput automatically accept
QDialogs. That was annoying, when more than one QLineEdit need to be
set.

[ChangeLog][Quick][Android] EnterKey type is now changed from
EnterKeyDefault to EnterKeyNext for virtual keyboard in TextInput. It is
done only if the focus can be moved to QQuickItem below, so the
activeFocusOnTab need to be set.

Fixes: QTBUG-61652
Change-Id: Ib2faf78f4c85788498476c1e67e67d29d508cfea
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>